move conditional test to a package
