+++ e2e/__tests__/testPathPatternReporterMessage.test.ts
@@ -3,7 +3,6 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- * @flow
  */
 
 'use strict';
