+++ packages/test-utils/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@jest/test-utils",
+  "version": "0.0.0",
+  "private": true,
+  "license": "MIT",
+  "main": "src/index.ts"
+}
